Refactor model organization.
